cloud_database/requirements_upload_ip.txt
netifaces>=0.10.6
python-firebase>=1.2